% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspect.R
\name{aspect}
\alias{aspect}
\title{Retrieve the `aspect` table from the milestones database}
\usage{
aspect()
}
\value{
Returns a tibble of the `aspect` table
}
\description{
The `aspect` table contains the classification of milestone events by their
role in the history of data visualization.
}
\details{
The fields in the `aspect` table are:
 \describe{
  \item{\code{asid}}{aspect id, a numeric vector}
  \item{\code{name}}{aspect name, a character vector with values:
    "Cartography", "Statistics & Graphics", "Technology", "Other"}
 }

 The aspect classification indicates the role an event played in the history
 of data visualization. Use \code{\link{milestone2aspect}} to link milestones
 to their aspects.
}
\examples{
# Get all aspects
asp <- aspect()

}
