% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{Open a MySQL connection to the milestones database}
\usage{
connect()
}
\value{
The string representing the database connection. For convenience, this is also
        stored in an environment variable, `.mstone.env$connnection`.
}
\description{
This version requires that you setup a keyring password for the source `milestones`
}
\examples{
\dontrun{
connect()
}

}
