% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html2latin1.R
\name{html2latin1}
\alias{html2latin1}
\title{Translate HTML entities to latin1 characters}
\usage{
html2latin1(txt)
}
\arguments{
\item{txt}{A vector of character strings}

\item{A}{vector of character strings#'}
}
\value{
A character vector, with HTML entities translated
}
\description{
This function should be ammended to handles a few milestone encodings that aren't quite correct in the DB.
}
\examples{
strings <- c("Fr&egrave;re de Montizon", "Lumi&egrave;re",
              "Ni&eacute;pce", "S&uuml;ssmilch", "Sch&uuml;pbach",
              "&#177; .25 &#215; 2 = &#189;")
html2latin1(strings)
}
\author{
David Carlson
}
