% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html2utf8.R
\name{html2utf8}
\alias{html2utf8}
\title{Translate HTML entities to UTF8 characters}
\usage{
html2utf8(text)
}
\arguments{
\item{text}{A vector of character strings}
}
\value{
A character vector, with HTML entities translated
}
\description{
This function is somewhat of a kludge to handle the encoding of various fields in the
`milestones` database in a form that there could be directly displayed in the PHP scripts.
}
\details{
In the present form, it uses \code{\link[textutils]{HTMLdecode}} for named HTML entities, like "&eacute;"
That doesn't handle numeric HTML entities, like "&#0239".
This function also handles a few milestone encodings that aren't quite correct in the DB.
}
\examples{
strings <- c("Fr&egrave;re de Montizon", "Lumi&egrave;re",
              "Ni&eacute;pce", "S&uuml;ssmilch", "Sch&uuml;pbach",
              "&#177; .25 &#215; 2 = &#189;")
html2utf8(strings)

}
