% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html2latin1.R
\name{latin12html}
\alias{latin12html}
\title{Translate latin1 characters to HTML entities}
\usage{
latin12html(txt, type = c("name", "number"))
}
\arguments{
\item{txt}{A vector of character strings with latin1 characters}

\item{type}{Type of characters to translate: \code{"name"} or \code{"number"}}
}
\value{
A character vector, with latin1 characters translated to HTML
}
\description{
Translate latin1 characters to HTML entities
}
\examples{
strings <- c("Frère de Montizon", "Lumière", "Niépce", "Süssmilch", "Schüpbach",
             "± .25 × 2 = ½")
latin12html(strings)
}
\author{
David Carlson
}
