% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaitem.R
\name{mediaitem}
\alias{mediaitem}
\title{Retrieve the `mediaitem` table from the milestones database}
\usage{
mediaitem()
}
\value{
Returns a tibble of the `mediaitem` table
}
\description{
The `mediaitem` table gives each link or image used in displaying the milestones items
}
\details{
The fields in the `mediaitem` table are:
\describe{
   \item{\code{miid}}{mediaitem id, a numeric vector}
   \item{\code{type}}{type, a character vector, with values \code{link}, \code{image}}
   \item{\code{url}}{URL to reference this item in the milestones application}
   \item{\code{title}}{a character vector}
   \item{\code{caption}}{a character vector}
   \item{\code{source}}{a character vector}
   \item{\code{mid}}{milestones id to which this is attached, a numeric vector}
}

For `type=="link"`, these give a standard HTML reference in the `url` field.
For `type=="image"`, the `url` field references the image in the milestones website.
}
