% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestone.R
\name{milestone}
\alias{milestone}
\title{Retrieve the `milestone` table from the milestones database}
\usage{
milestone()
}
\value{
Returns a tibble of the `milestone` table
}
\description{
The `milestone` table gives each item considered a milestone in the history of data visualization.
}
