% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestone2author.R
\name{milestone2author}
\alias{milestone2author}
\alias{milestone2author-data}
\title{Retrieve the `milestone2author` table from the milestones database}
\format{
A data frame. See \code{\link{milestone2author}} for details.
}
\usage{
milestone2author()
}
\value{
Returns a tibble of the `milestone2author` table
}
\description{
The `milestone2author` table links milestones to their authors.
}
\details{
The fields in the `milestone2author` table are:
 \describe{
  \item{\code{mid}}{milestone id, a numeric vector}
  \item{\code{aid}}{author id, a numeric vector}
 }

 This table links milestones to authors. Use with \code{\link{authors}}
 to get author information.
}
\examples{
\dontrun{
# Get milestone-author links
m2a <- milestone2author()
}

}
\seealso{
\code{\link{authors}}, \code{\link{milestone}}
}
