% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subject.R
\docType{data}
\name{subject}
\alias{subject}
\title{Retrieve the `subject` table from the milestones database}
\format{
A data frame with 4 rows and 2 variables:
\describe{
  \item{sid}{subject id, a numeric vector}
  \item{name}{subject name (Physical, Mathematical, Human, Other), a character vector}
}
}
\source{
\url{https://datavis.ca/milestones/}
}
\usage{
subject()

subject
}
\value{
Returns a tibble of the `subject` table
}
\description{
The `subject` table contains the classification of milestone events by their
substantive context.

A lookup table containing subject classifications that indicate the
substantive context of milestone events.
}
\details{
The fields in the `subject` table are:
 \describe{
  \item{\code{sid}}{subject id, a numeric vector}
  \item{\code{name}}{subject name, a character vector with values:
    "Physical", "Mathematical", "Human", "Other"}
 }

 The subject classification indicates the substantive context of the milestone event.
 Use \code{\link{milestone2subject}} to link milestones to their subjects.
}
\examples{
# Get all subjects
subj <- subject()

}
\seealso{
\code{\link{milestone2subject}}
}
\keyword{datasets}
