% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspect.R
\name{aspect}
\alias{aspect}
\alias{aspect-data}
\title{Aspect classification data from the Milestones Project}
\format{
A data frame. See \code{\link{aspect}} for details.
}
\source{
\url{https://datavis.ca/milestones/}
}
\usage{
aspect()
}
\value{
A data frame with 4 rows and 2 variables:
\describe{
  \item{asid}{aspect id, a numeric vector}
  \item{name}{aspect name (Cartography, Statistics & Graphics, Technology, Other), a character vector}
}
}
\description{
Retrieve the `aspect` table containing aspect classifications that indicate
the role milestone events played in the history of data visualization.
}
\details{
The aspect classification indicates the role an event played in the history
 of data visualization. Use \code{\link{milestone2aspect}} to link milestones
 to their aspects.
}
\examples{
\dontrun{
# Get all aspects
aspect()
}

}
\seealso{
\code{\link{milestone2aspect}}
}
