% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/author.R
\name{authors}
\alias{authors}
\title{Retrieve the `author` table from the milestones database}
\usage{
authors()
}
\value{
Returns a tibble of the `author` table
}
\description{
Retrieve the `author` table from the milestones database
}
\details{
The `author` data set contains the following fields:

  \describe{
     \item{\code{aid}}{author id, a numeric vector}
     \item{\code{prefix}}{name prefix, a character vector}
     \item{\code{givennames}}{givennames, a character vector}
     \item{\code{lname}}{last name, a character vector}
     \item{\code{suffix}}{name suffix, a character vector}
     \item{\code{lived}}{years lived, a character vector}
     \item{\code{birthdate}}{a character vector}
     \item{\code{birthplace}}{place of birth, a character vector}
     \item{\code{birthlat}}{latitude of birthplace, a numeric vector}
     \item{\code{birthlong}}{longitude of birthplace, a numeric vector}
     \item{\code{deathdate}}{a character vector}
     \item{\code{deathplace}}{a character vector}
     \item{\code{deathlat}}{latitude of deathplace, a numeric vector}
     \item{\code{deathlong}}{longitude of deathplace, a numeric vector}
     \item{\code{note}}{notes about this author, a character vector}
   }

TODO: The various names field contain some HTML characters.  These should be
translated to latin1.
}
\examples{
# none yet

}
