% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword.R
\name{keyword}
\alias{keyword}
\alias{keyword-data}
\title{Keyword data from the Milestones Project}
\format{
A data frame. See \code{\link[=keyword]{keyword()}} for details.
}
\source{
\url{https://datavis.ca/milestones/}
}
\usage{
keyword()
}
\value{
A data frame with 335 rows and 2 variables:
\describe{
\item{kid}{keyword id, a numeric vector}
\item{name}{keyword name, a character vector}
}
}
\description{
Retrieve the \code{keyword} table containing keywords or terms attached to
milestone items as a freeform classification.
}
\details{
This is a freeform classification listing keywords or terms attached to
milestone items. Use \code{\link[=milestone2keyword]{milestone2keyword()}} to link milestones
to their keywords.
}
\examples{
\dontrun{
# Get all keywords
kw <- keyword()

# Find keywords containing "chart"
library(dplyr)
kw |> filter(grepl("chart", name, ignore.case = TRUE))
}

}
\seealso{
\code{\link[=milestone2keyword]{milestone2keyword()}}
}
