% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestone.R
\name{milestone}
\alias{milestone}
\alias{milestone-data}
\title{Milestone data from the Milestones Project}
\format{
A data frame. See \code{\link[=milestone]{milestone()}} for details.
}
\source{
\url{https://datavis.ca/milestones/}
}
\usage{
milestone()
}
\value{
A data frame with 297 rows and 14 variables:
\describe{
\item{mid}{milestone id, a numeric vector}
\item{slug}{title slug, a character vector}
\item{date_from_numeric}{numeric start date, a numeric vector}
\item{date_from}{start date text, a character vector}
\item{date_to_numeric}{numeric end date, a numeric vector}
\item{date_to}{end date text, a character vector}
\item{tag}{tag line, a character vector}
\item{description}{description of the milestone, a character vector}
\item{location}{location, a character vector}
\item{add_date}{date added to database, a character vector}
\item{modified_date}{date last modified, a character vector}
\item{status}{publication status, a character vector}
\item{note}{notes, a character vector}
\item{extra}{extra information, a character vector}
}
}
\description{
Retrieve the \code{milestone} table containing milestone events in the history of
data visualization. This is the primary table in the database. Other tables
are linked to it via the \code{mid} key.
}
\examples{
\dontrun{
# Get all milestones
m <- milestone()

# Filter by date
library(dplyr)
m |> filter(date_from_numeric < 0)  # BC milestones
}

}
