% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestone2author.R
\docType{data}
\name{milestone2author}
\alias{milestone2author}
\title{Retrieve the `milestone2author` table from the milestones database}
\format{
A data frame with 348 rows and 2 variables:
\describe{
  \item{mid}{milestone id, a numeric vector}
  \item{aid}{author id, a numeric vector}
}
}
\source{
\url{https://datavis.ca/milestones/}
}
\usage{
milestone2author()

milestone2author
}
\value{
Returns a tibble of the `milestone2author` table
}
\description{
The `milestone2author` table links milestones to their authors.

A linking table that associates milestones with their authors.
}
\details{
The fields in the `milestone2author` table are:
 \describe{
  \item{\code{mid}}{milestone id, a numeric vector}
  \item{\code{aid}}{author id, a numeric vector}
 }

 This table links milestones to authors. Use with \code{\link{authors}}
 to get author information.
}
\examples{
# Get milestone-author links
m2a <- milestone2author()

}
\seealso{
\code{\link{authors}}, \code{\link{milestone}}

\code{\link{authors}}, \code{\link{milestone}}
}
\keyword{datasets}
