% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestone2reference.R
\docType{data}
\name{milestone2reference}
\alias{milestone2reference}
\title{Retrieve the `milestone2reference` table from the milestones database}
\format{
A data frame with 347 rows and 3 variables:
\describe{
  \item{mid}{milestone id, a numeric vector}
  \item{rid}{reference id, a numeric vector}
  \item{extra}{extra information about the reference, a character vector}
}
}
\source{
\url{https://datavis.ca/milestones/}
}
\usage{
milestone2reference()

milestone2reference
}
\value{
Returns a tibble of the `milestone2reference` table
}
\description{
The `milestone2reference` table links milestones to their references.

A linking table that associates milestones with their bibliographic references.
}
\details{
The fields in the `milestone2reference` table are:
 \describe{
  \item{\code{mid}}{milestone id, a numeric vector}
  \item{\code{rid}}{reference id, a numeric vector}
 }

 This table links milestones to references. Use with \code{\link{reference}}
 to get reference information.
}
\examples{
# Get milestone-reference links
m2r <- milestone2reference()

}
\seealso{
\code{\link{reference}}, \code{\link{milestone}}

\code{\link{reference}}, \code{\link{milestone}}
}
\keyword{datasets}
