% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestone2reference.R
\name{milestone2reference}
\alias{milestone2reference}
\alias{milestone2reference-data}
\title{Retrieve the `milestone2reference` table from the milestones database}
\format{
A data frame. See \code{\link{milestone2reference}} for details.
}
\usage{
milestone2reference()
}
\value{
Returns a tibble of the `milestone2reference` table
}
\description{
The `milestone2reference` table links milestones to their references.
}
\details{
The fields in the `milestone2reference` table are:
 \describe{
  \item{\code{mid}}{milestone id, a numeric vector}
  \item{\code{rid}}{reference id, a numeric vector}
 }

 This table links milestones to references. Use with \code{\link{reference}}
 to get reference information.
}
\examples{
\dontrun{
# Get milestone-reference links
m2r <- milestone2reference()
}

}
\seealso{
\code{\link{reference}}, \code{\link{milestone}}
}
