% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestone2subject.R
\name{milestone2subject}
\alias{milestone2subject}
\alias{milestone2subject-data}
\title{Retrieve the `milestone2subject` table from the milestones database}
\format{
A data frame. See \code{\link{milestone2subject}} for details.
}
\usage{
milestone2subject()
}
\value{
Returns a tibble of the `milestone2subject` table
}
\description{
The `milestone2subject` table links milestones to their subject classifications.
}
\details{
The fields in the `milestone2subject` table are:
 \describe{
  \item{\code{mid}}{milestone id, a numeric vector}
  \item{\code{sid}}{subject id, a numeric vector}
  \item{\code{subject}}{subject name, a character vector}
 }

 This table links milestones to subjects. The \code{subject} field is included
 for convenience (joined from the \code{\link{subject}} table).
}
\examples{
\dontrun{
# Get milestone-subject links
m2s <- milestone2subject()
}

}
\seealso{
\code{\link{subject}}, \code{\link{milestone}}
}
