% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subject.R
\name{subject}
\alias{subject}
\title{Retrieve the `subject` table from the milestones database}
\usage{
subject()
}
\value{
Returns a tibble of the `subject` table
}
\description{
The `subject` table contains the classification of milestone events by their
substantive context.
}
\details{
The fields in the `subject` table are:
 \describe{
  \item{\code{sid}}{subject id, a numeric vector}
  \item{\code{name}}{subject name, a character vector with values:
    "Physical", "Mathematical", "Human", "Other"}
 }

 The subject classification indicates the substantive context of the milestone event.
 Use \code{\link{milestone2subject}} to link milestones to their subjects.
}
\examples{
# Get all subjects
subj <- subject()

}
