% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subject.R
\name{subject}
\alias{subject}
\alias{subject-data}
\title{Subject classification data from the Milestones Project}
\format{
A data frame. See \code{\link[=subject]{subject()}} for details.
}
\source{
\url{https://datavis.ca/milestones/}
}
\usage{
subject()
}
\value{
A data frame with 4 rows and 2 variables:
\describe{
\item{sid}{subject id, a numeric vector}
\item{name}{subject name (Physical, Mathematical, Human, Other), a character vector}
}
}
\description{
Retrieve the \code{subject} table containing subject classifications that indicate
the substantive context of milestone events.
}
\details{
The subject classification indicates the substantive context of the milestone event.
Use \code{\link[=milestone2subject]{milestone2subject()}} to link milestones to their subjects.
}
\examples{
\dontrun{
# Get all subjects
subj <- subject()
}

}
\seealso{
\code{\link[=milestone2subject]{milestone2subject()}}
}
